+++ GapsWeb/src/main/java/com/jasonhhouse/gaps/SearchCancelledException.java
@@ -8,37 +8,25 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.jasonhhouse.Gaps;
+package com.jasonhhouse.gaps;
 
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-/**
- * Search for all missing movies in your plex collection by MovieDB collection.
- */
-@SpringBootApplication
-@EnableAsync
-public class GapsApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(GapsApplication.class, args);
+public class SearchCancelledException extends Exception {
+    public SearchCancelledException() {
     }
 
-    @Bean
-    public Executor taskExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(2);
-        executor.setMaxPoolSize(2);
-        executor.setQueueCapacity(500);
-        executor.setThreadNamePrefix("Gaps-");
-        executor.initialize();
-        return executor;
+    public SearchCancelledException(String message) {
+        super(message);
     }
 
+    public SearchCancelledException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public SearchCancelledException(Throwable cause) {
+        super(cause);
+    }
+
+    public SearchCancelledException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
 }
