+++ GapsWeb/src/main/resources/application.yaml
@@ -44,3 +44,9 @@ server:
     key-alias: gaps
 security:
   require-ssl: true
+
+info:
+  app:
+    name: Gaps
+    description: Gaps searches through your Plex Server or local folders for all movies, then queries for known movies in the same collection. If those movies don't exist in your library, Gaps will recommend getting those movies, legally of course.
+    version: 0.0.5
